Rename pkg name to externalerror
